requirements.txt
django==4.1.7
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg-binary==3.1.8
psycopg2==2.9.5